+++ __tests__/index.ts
@@ -2,6 +2,7 @@ import { memcache, spanner } from '@pulumi/gcp'
 
 export * as artifactregistry from './artifactregistry'
 export * as biguery from './bigquery'
+export * as bigtable from './bigtable'
 export * as cloudfunctions from './cloudfunctions'
 export * as cloudfunctionsv2 from './cloudfunctionsv2'
 export * as cloudrun from './cloudrun'