+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Bigtable policies [#39](https://github.com/losisin/gcp-pac/pull/39)
+  - `requireDeletionProtection`
+  - `requireInstanceCmek`
+
 ## [1.7.0] - 2024-01-22
 
 ### Added